+++ core/services/eth/errors.go
@@ -96,7 +96,11 @@ var optimism = ClientErrors{
 	FeeTooHigh: regexp.MustCompile(`(: |^)fee too high: \d+, use less than \d+ \* [0-9\.]+$`),
 }
 
-var clients = []ClientErrors{parity, geth, arbitrum, optimism}
+var avalanche = ClientErrors{
+	NonceTooLow: regexp.MustCompile(`(: |^)nonce too low: address 0x[0-9a-fA-F]{40} current nonce \([\d]+\) > tx nonce \([\d]+\)$`),
+}
+
+var clients = []ClientErrors{parity, geth, arbitrum, optimism, avalanche}
 
 func (s *SendError) is(errorType int) bool {
 	if s == nil || s.err == nil {